Refactor recent projects into component, use all available projects
